refacto: restructure MCP server
