+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
 	"crates/sozo/scarbext",
 	"crates/sozo/signers",
 	"crates/sozo/walnut",
+	"crates/sozo/mcp",
 	"examples/spawn-and-move",
 	"scripts/verify_db_balances",
 	"xtask/generate-test-db",
@@ -79,6 +80,7 @@ sozo-ops = { path = "crates/sozo/ops" }
 sozo-scarbext = { path = "crates/sozo/scarbext" }
 sozo-signers = { path = "crates/sozo/signers" }
 sozo-walnut = { path = "crates/sozo/walnut" }
+sozo-mcp = { path = "crates/sozo/mcp" }
 
 # macros
 merge-options = { path = "crates/macros/merge-options" }
